fix: login modal not visible in small screen
