+++ packages/hoppscotch-common/src/components/app/ActionHandler.vue
@@ -1,0 +1,26 @@
+<template>
+  <AppShortcuts :show="showShortcuts" @close="showShortcuts = false" />
+  <AppShare :show="showShare" @hide-modal="showShare = false" />
+  <FirebaseLogin :show="showLogin" @hide-modal="showLogin = false" />
+</template>
+
+<script setup lang="ts">
+import { ref } from "vue"
+import { defineActionHandler } from "~/helpers/actions"
+
+const showShortcuts = ref(false)
+const showShare = ref(false)
+const showLogin = ref(false)
+
+defineActionHandler("flyouts.keybinds.toggle", () => {
+  showShortcuts.value = !showShortcuts.value
+})
+
+defineActionHandler("modals.share.toggle", () => {
+  showShare.value = !showShare.value
+})
+
+defineActionHandler("modals.login.toggle", () => {
+  showLogin.value = !showLogin.value
+})
+</script>
